+++ example/lib/main.dart
@@ -1,6 +1,17 @@
 import 'package:flutter/material.dart';
 import 'package:skeleton_animation/skeleton_animation.dart';
 
+List tileList = [
+  [null, null],
+  [null, Colors.blue],
+  [null, Colors.blue.shade900],
+  [Colors.green, Colors.blue.shade900],
+  [null, Colors.green],
+  [null, Colors.green.shade900],
+  [null, Colors.red],
+  [null, Colors.red.shade200]
+];
+
 void main() {
   runApp(MyApp());
 }
@@ -14,26 +25,29 @@ class MyApp extends StatelessWidget {
           title: Text('Skeleton animation'),
         ),
         body: Container(
-            padding: EdgeInsets.symmetric(horizontal: 16.0),
             child: CustomScrollView(
-              slivers: [
-                SliverList(
-                  delegate: SliverChildBuilderDelegate(
-                      (context, i) => ListTile(
-                            title: Skeleton(
-                              style: SkeletonStyle.text,
-                              height: 14.0,
-                            ),
-                            subtitle: Skeleton(
-                              style: SkeletonStyle.text,
-                              height: 10.0,
-                            ),
-                            onTap: () {},
-                          ),
-                      childCount: 20),
-                )
-              ],
-            )),
+          slivers: [
+            SliverList(
+              delegate: SliverChildBuilderDelegate(
+                  (context, i) => ListTile(
+                        contentPadding: EdgeInsets.symmetric(horizontal: 16.0),
+                        tileColor: tileList[i][1],
+                        title: Skeleton(
+                          style: SkeletonStyle.text,
+                          textColor: tileList[i][0],
+                          parentBackgroundColor: tileList[i][1],
+                          height: 14.0,
+                        ),
+                        subtitle: Text(
+                          'Loaded Text',
+                          style: TextStyle(color: tileList[i][0]),
+                        ),
+                        onTap: () {},
+                      ),
+                  childCount: tileList.length),
+            )
+          ],
+        )),
       ),
       // enable dark mode
       darkTheme: ThemeData(brightness: Brightness.dark),