+++ CHANGELOG.md
@@ -1,5 +1,9 @@
-## [0.1.0] - Border support
+## [0.1.0] - Background color fix / Border support
 
+* Added a option to set the parent background color
+* Removed wave animation
+* Removed hightlightColor
+* Changed baseColor to textColor
 * Added support for borders
 
 ## [0.0.3] - Fixed issues